requirements.txt
# API
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-multipart>=0.0.6
jinja2>=3.1.2

# Data processing
pandas>=1.5.3
numpy>=1.24.2
scikit-learn>=1.2.2
xgboost>=1.7.5
lightgbm>=3.3.5
catboost>=1.2
joblib>=1.2.0
pyyaml>=6.0

# AWS
boto3>=1.26.115
botocore>=1.29.115

# Visualization
matplotlib>=3.7.1
seaborn>=0.12.2
plotly>=5.14.1

# Optimization
optuna>=3.1.1

# Testing
pytest>=7.3.1
httpx>=0.24.0

# Development
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.2.0
